+++ examples/platforms/android/app/src/test/java/com/here/android/test/ArraysByteBufferTest.java
@@ -61,9 +61,9 @@ public class ArraysByteBufferTest {
 
   @Test
   public void methodWithImplicitArray_emptyArray() {
-    List<Long> uintList = new ArrayList<>();
+    List<Short> uintList = new ArrayList<>();
 
-    List<Long> resultList = ArraysByteBuffer.methodWithImplicitArray(uintList);
+    List<Short> resultList = ArraysByteBuffer.methodWithImplicitArray(uintList);
 
     assertNotNull(resultList);
     assertEquals(0, resultList.size());
@@ -71,19 +71,19 @@ public class ArraysByteBufferTest {
 
   @Test
   public void methodWithImplicitArray_reversesArray() {
-    List<Long> uintList = java.util.Arrays.asList(1L, 2L, 3L);
+    List<Short> uintList = java.util.Arrays.asList((short)1, (short)2, (short)3);
 
-    List<Long> resultList = ArraysByteBuffer.methodWithImplicitArray(uintList);
+    List<Short> resultList = ArraysByteBuffer.methodWithImplicitArray(uintList);
 
     assertNotNull(resultList);
-    assertEquals(java.util.Arrays.asList(3L, 2L, 1L), resultList);
+    assertEquals(java.util.Arrays.asList((short)3, (short)2, (short)1), resultList);
   }
 
   @Test
   public void methodWithExplicitArray_emptyArray() {
-    List<Long> uintList = new ArrayList<>();
+    List<Short> uintList = new ArrayList<>();
 
-    List<Long> resultList = ArraysByteBuffer.methodWithExplicitArray(uintList);
+    List<Short> resultList = ArraysByteBuffer.methodWithExplicitArray(uintList);
 
     assertNotNull(resultList);
     assertEquals(0, resultList.size());
@@ -91,12 +91,12 @@ public class ArraysByteBufferTest {
 
   @Test
   public void methodWithExplicitArray_reversesArray() {
-    List<Long> uintList = java.util.Arrays.asList(1L, 2L, 3L);
+    List<Short> uintList = java.util.Arrays.asList((short)1, (short)2, (short)3);
 
-    List<Long> resultList = ArraysByteBuffer.methodWithExplicitArray(uintList);
+    List<Short> resultList = ArraysByteBuffer.methodWithExplicitArray(uintList);
 
     assertNotNull(resultList);
-    assertEquals(java.util.Arrays.asList(3L, 2L, 1L), resultList);
+    assertEquals(java.util.Arrays.asList((short)3, (short)2, (short)1), resultList);
   }
 
   @Test
@@ -140,13 +140,13 @@ public class ArraysByteBufferTest {
   public void methodWithImplicitArrayInStruct_reversesArray() {
     ArraysByteBuffer.StructWithImplicitArray struct =
         new ArraysByteBuffer.StructWithImplicitArray();
-    struct.image = java.util.Arrays.asList(1L, 2L, 3L);
+    struct.image = java.util.Arrays.asList((short)1, (short)2, (short)3);
 
     ArraysByteBuffer.StructWithImplicitArray resultStruct =
         ArraysByteBuffer.methodWithImplicitArrayInStruct(struct);
 
     assertNotNull(resultStruct);
-    assertEquals(java.util.Arrays.asList(3L, 2L, 1L), resultStruct.image);
+    assertEquals(java.util.Arrays.asList((short)3, (short)2, (short)1), resultStruct.image);
   }
 
   @Test
@@ -166,12 +166,12 @@ public class ArraysByteBufferTest {
   public void methodWithExplicitArrayInStruct_reversesArray() {
     ArraysByteBuffer.StructWithExplicitArray struct =
         new ArraysByteBuffer.StructWithExplicitArray();
-    struct.image = java.util.Arrays.asList(1L, 2L, 3L);
+    struct.image = java.util.Arrays.asList((short)1, (short)2, (short)3);
 
     ArraysByteBuffer.StructWithExplicitArray resultStruct =
         ArraysByteBuffer.methodWithExplicitArrayInStruct(struct);
 
     assertNotNull(resultStruct);
-    assertEquals(java.util.Arrays.asList(3L, 2L, 1L), resultStruct.image);
+    assertEquals(java.util.Arrays.asList((short)3, (short)2, (short)1), resultStruct.image);
   }
 }